+++ controllers/contacts/getContactById.js
@@ -1,0 +1,12 @@
+import { HttpError } from "../../helpers/index.js";
+import { ctrlWrapper } from "../../decorators/index.js";
+import { Contact } from "../../models/contact.js";
+
+const getById = async (req, res) => {
+  const { id } = req.params;
+  const contact = await Contact.findById(id);
+  if (!contact) throw HttpError(404);
+  res.json(contact);
+};
+
+export const getContactById = ctrlWrapper(getById);
