+++ controllers/contacts/index.js
@@ -1,0 +1,6 @@
+export * from "./getAllContacts.js";
+export * from "./getContactById.js";
+export * from "./addContact.js";
+export * from "./deleteContactById.js";
+export * from "./updateContactById.js";
+export * from "./updateStatusContact.js";
